requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.1
    # via starlette
fastapi==0.78.0
    # via -r requirements.in
idna==3.3
    # via anyio
pydantic==1.9.1
    # via fastapi
python-dotenv==0.20.0
    # via -r requirements.in
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
typing-extensions==4.2.0
    # via
    #   pydantic
    #   starlette
